Put getDecimalPlaces in utilities
