+++ packages/components/src/components/as-range-slider/slider/as-range-slider.tsx
@@ -1,4 +1,5 @@
 import { Component, Event, EventEmitter, Prop, State, Watch } from '@stencil/core';
+import getDecimalPlaces from '../../../utils/get-decimal-places';
 import { Thumb } from '../thumb/as-range-slider-thumb';
 
 @Component({
@@ -309,20 +310,4 @@ export class RangeSlider {
   private roundToStep(numberToRound: number, step: number) {
     return Number.parseFloat(numberToRound.toFixed(getDecimalPlaces(step)));
   }
-}
-
-function getDecimalPlaces(decimalNumber): number {
-  // Copied this method from: https://bit.ly/2DfxbfQ
-
-  function hasFraction(numberToCheck) {
-    return Math.abs(Math.round(numberToCheck) - numberToCheck) > 1e-10;
-  }
-
-  let count = 0;
-  // multiply by increasing powers of 10 until the fractional part is ~ 0
-  while (hasFraction(decimalNumber * (10 ** count)) && isFinite(10 ** count)) {
-    count++;
-  }
-
-  return count;
 }
